requirements.txt
spec
nose
pep8
behave
coverage
scrutinizer-ocular
pbkdf2
pycrypto
jinja2
